Various rport-plus related tidyups
