+++ rport-plus/capabilities/oauth/oauth.go
@@ -42,6 +42,11 @@ type Config struct {
 	ProvideOAuthLogin    bool   `mapstructure:"provide_oauth_login"`
 }
 
+const (
+	InitOAuthCapabilityEx = "InitOAuthCapabilityEx"
+	GitHubOAuthProvider   = "github"
+)
+
 // Capability is used by rportd to maintain loaded info about the plugin's
 // oauth capability
 type Capability struct {
@@ -51,24 +56,26 @@ type Capability struct {
 	Logger *logger.Logger
 }
 
-const (
-	InitOAuthCapabilityEx = "InitOAuthCapabilityEx"
-	GitHubOAuthProvider   = "github"
-)
-
+// GetInitFuncName gets the name of the capability init func
 func (cap *Capability) GetInitFuncName() (name string) {
 	return InitOAuthCapabilityEx
 }
 
+// SetProvider invokes the capability init func in the plugin and saves
+// the returned capability provider interface. This interface provides
+// the functions of the capability.
 func (cap *Capability) SetProvider(initFn plugin.Symbol) {
 	fn := initFn.(func(cap *Capability) (capProvider CapabilityEx))
 	cap.Provider = fn(cap)
 }
 
+// GetOAuthCapabilityEx returns the interface to the capability functions
 func (cap *Capability) GetOAuthCapabilityEx() (capEx CapabilityEx) {
 	return cap.Provider
 }
 
+// GetConfigValidator returns a validator interface that can be called to
+// validate the capability config
 func (cap *Capability) GetConfigValidator() (v validator.Validator) {
 	return cap.Provider
 }
