--- rport-plus/symbols/symbols_test.go
@@ -1,7 +1,0 @@
-package symbols_test
-
-import "testing"
-
-func TestShould(t *testing.T) {
-
-}
