+++ rport-plus/validator/validator.go
@@ -1,6 +1,7 @@
 package validator
 
-// Validator provides a narrow interface for capability validation
+// Validator provides a narrow interface for capability validation.
+// In a separate package to avoid circular dependencies.
 type Validator interface {
 	ValidateConfig() (err error)
 }
